refactor(sql-orm-client): rename query-plan-scope.ts to query-plan-source.ts
The identifier rename landed, but the filename still said "scope" -
the last place that word survived in this meaning, and the first
thing a reader meets. Everything inside the module is
buildAggregateInput, buildDedupedTableSource, buildStateWhere,
buildMtiJoins, wrapWithRowNumberDedup, createTableRefRemapper - it
owns the FROM/WHERE/JOIN side of a plan (which rows, and where they
come from), as opposed to query-plan-select.ts (how a read is shaped)
or query-plan-aggregate.ts (how rows are reduced).

"Source" matches this codebase's existing vocabulary (AnyFromSource,
TableSource, DerivedTableSource - two of the module's six exports
return one). query-plan-rows.ts was considered and rejected: already
taken by the queryPlanRows execution helper.

git mv, so history follows the file. Two import path updates
(query-plan-select.ts, query-plan-aggregate.ts) - the only non-transient
references repo-wide; everything else naming the old path lives under
projects/aggregate-pagination/**, which stays as-is. No module-level
doc comment existed to update. Zero behaviour change: every test
passes with the same pass/fail counts as before this commit.
